service/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.14.0
filelock==3.8.0
idna==3.4
platformdirs==2.5.4
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
virtualenv==20.16.7
